Se actualizo y refactorizo el codigo
