+++ package.json
@@ -7,8 +7,8 @@
     "start": "node server.js",
     "dev-start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "browserify public/js/game.js -o public/js/bundle.js",
-    "watch": "watchify public/js/game.js -o public/js/bundle.js"
+    "build": "webpack --config webpack.config.js",
+    "watch": "webpack --watch --config webpack.config.js"
   },
   "repository": {
     "type": "git",
