+++ server.js
@@ -2,10 +2,11 @@ const express = require("express");
 const path = require("path");
 const app = express();
 
-app.use(express.static(path.join(__dirname, '/public')));
+app.use('/dist', express.static(path.join(__dirname, '/dist')));
+app.use('/assets', express.static(path.join(__dirname, '/assets')));
 
 app.get('/', (req, res) => {
-    res.sendFile(__dirname + '/index.html');
+    res.sendFile(__dirname + '/dist/index.html');
 });
 
 app.listen(8080, () => {console.log("Game Server Started at http://localhost:8080")});
